fix: product images into combo included

+++ src/modules/combo/infrastructure/combo.service.ts
@@ -33,9 +33,17 @@ export class ComboService {
   }
 
   async findAll() {
-    return this.comboRepository.find({
+    const combos = await this.comboRepository.find({
       relations: ['products', 'products.images'],
     });
+
+    return combos.map(combo => ({
+      ...combo,
+      products: combo.products.map(product => ({
+        ...product,
+        images: product.images.map(img => img.image_url),
+      })),
+    }));
   }
 
   async findOne(id: number) {
@@ -48,7 +56,13 @@ export class ComboService {
       throw new NotFoundException(`Combo with id ${id} not found`);
     }
 
-    return combo;
+    return {
+      ...combo,
+      products: combo.products.map(product => ({
+        ...product,
+        images: product.images.map(img => img.image_url),
+      })),
+    };
   }
 
   async update(id: number, updateComboDto: UpdateComboDto) {
@@ -73,7 +87,10 @@ export class ComboService {
   }
 
   async remove(id: number) {
-    const combo = await this.findOne(id);
+    const combo = await this.comboRepository.findOne({ where: { combo_id: id } });
+    if (!combo) {
+      throw new NotFoundException(`Combo with id ${id} not found`);
+    }
     await this.comboRepository.remove(combo);
   }
 }
